CRUD functionality with Firestore support

+++ app/src/main/java/dk/rskovbo/md_android_notebook/NoteActivity.kt
@@ -1,42 +1,87 @@
 package dk.rskovbo.md_android_notebook
 
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
+import android.view.Menu
+import android.view.MenuItem
 import android.widget.EditText
+import android.widget.Toast
+import dk.rskovbo.md_android_notebook.MainActivity.Companion.db
+import dk.rskovbo.md_android_notebook.MainActivity.Companion.noteItems
 
 class NoteActivity : AppCompatActivity() {
+    lateinit var editTitle: EditText
+    lateinit var editBody: EditText
+    lateinit var builder: AlertDialog.Builder
+    var title: String? = ""
+    var body: String? = ""
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_note)
 
-        val title = intent.extras?.getString(TITLE)
-        val body = intent.extras?.getString(BODY)
-        val editText: EditText = findViewById(R.id.editText)
-        setTitle(title)
-        editText.setText(body)
+        title = intent.extras?.getString(TITLE)
+        body = intent.extras?.getString(BODY)
 
-        findViewById<Button>(R.id.saveButton).setOnClickListener{
-            saveBody(editText.text.toString())
-        }
+        editBody = findViewById(R.id.editBody)
+        editTitle = findViewById(R.id.editTitle)
 
-        findViewById<Button>(R.id.deleteButton).setOnClickListener{
+        setTitle("Notebook")
+        editTitle.setText(title)
+        editBody.setText(body)
+
+        builder = AlertDialog.Builder(this)
+        builder.setTitle("Confirm deletion")
+        builder.setMessage("Please confirm deletion.")
+        builder.setPositiveButton("OK") { dialog, which ->
             deleteItemInNote(this)
         }
+        builder.setNegativeButton("Cancel") { dialog, which ->
+            Toast.makeText(applicationContext, "Deletion cancelled", Toast.LENGTH_SHORT).show()
+        }
+
+        setSupportActionBar(findViewById(R.id.top_toolbar))
+
+    }
+
+    // Connect menu layout to toolbar
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.show_note_toolbar, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
+        R.id.save_note -> {
+            saveItem(editTitle.text.toString(), editBody.text.toString())
+            true
+        }
+        R.id.delete_note -> {
+            builder.show()
+            false
+        }
+        else -> super.onOptionsItemSelected(item)
     }
 
     // TODO Move to main activity and refactor
-    fun saveBody(editedBody: String) {
-        MainActivity.noteItems.get(MainActivity.currentIndex).body = editedBody
+    fun saveItem(editedTitle: String, editedBody: String) {
+        val currentIndex = MainActivity.currentIndex
+        val currentNote = noteItems.get(currentIndex)
+
+        val item = MovieItem(editedTitle, editedBody, currentNote.noteId)
+        noteItems.set(currentIndex, item)
+
+        db.collection("notes").document(item.noteId).set(item)
+
+        MainActivity.adapter.notifyDataSetChanged()
     }
 
     fun deleteItemInNote(context: Context) {
         val position = MainActivity.currentIndex
         ListAdapter.deleteItem(position)
         finish()
-        //startActivity(Intent(context, MainActivity::class.java))
     }
 
     companion object {
